Agenda implemented, many hacks reduced; Blocks World works, but --split-min/max > 3 results in reduced performance; Dynamic refinement seems to have regressed in Puddle World (performance for min 7 max 9 is worse than min/max of 7);

+++ src/carli/node.cpp
@@ -3,8 +3,10 @@
 #include "agent.h"
 
 Node::~Node() {
-  agent.purge_q_value(q_value);
-  q_value.delete_and_zero();
+  if(delete_q_value) {
+    agent.purge_q_value(q_value);
+    q_value.delete_and_zero();
+  }
 }
 
 void Node::destroy() {
@@ -17,10 +19,18 @@ Node_Split::Node_Split(Agent &agent_, const tracked_ptr<Q_Value> &q_value_)
   ++agent.q_value_count;
 }
 
+Node_Split::~Node_Split() {
+  --agent.q_value_count;
+}
+
 Node_Unsplit::Node_Unsplit(Agent &agent_, const size_t &depth_)
  : Node(agent_, new Q_Value(0.0, Q_Value::Type::UNSPLIT, depth_))
 {
   ++agent.q_value_count;
+}
+
+Node_Unsplit::~Node_Unsplit() {
+  --agent.q_value_count;
 }
 
 Node_Fringe::~Node_Fringe() {
